requirements.txt
requests
openai
python-dotenv
PyPDF2
google-generativeai
singlestoredb
